Allow either element ID or name to be specified + better null or default handling.

+++ HR.PersonalCalendar.WebApi/Extensions/PersonalizationExtensions.cs
@@ -12,11 +12,11 @@ namespace HR.PersonalCalendar.WebApi.Extensions
         {
             return personalization.ElementType switch
             {
-                ElementType.Klasse => new Klasse { Id = personalization.ElementId, Name = personalization.ElementName },
-                ElementType.Teacher => new Teacher { Id = personalization.ElementId, Name = personalization.ElementName },
-                ElementType.Subject => new Subject { Id = personalization.ElementId, Name = personalization.ElementName },
-                ElementType.Room => new Room { Id = personalization.ElementId, Name = personalization.ElementName },
-                ElementType.Student => new Student { Id = personalization.ElementId, Name = personalization.ElementName },
+                ElementType.Klasse => new Klasse { Id = (int)personalization.ElementId, Name = personalization.ElementName },
+                ElementType.Teacher => new Teacher { Id = (int)personalization.ElementId, Name = personalization.ElementName },
+                ElementType.Subject => new Subject { Id = (int)personalization.ElementId, Name = personalization.ElementName },
+                ElementType.Room => new Room { Id = (int)personalization.ElementId, Name = personalization.ElementName },
+                ElementType.Student => new Student { Id = (int)personalization.ElementId, Name = personalization.ElementName },
                 _ => throw new InvalidEnumArgumentException($"{nameof(personalization)}.{nameof(personalization.ElementType)}", Convert.ToInt32(personalization.ElementType), typeof(ElementType))
             };
         }
